+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2024-03-29
+## [1.1.0] - 2025-03-30
+
+### Added
+
+- Support for code blocks (inline and multi-line)
+- Support for strikethrough text
+- Support for links
+- Support for images (embedded, non-clickable)
+
+## [1.0.0] - 2025-03-29
 
 ### Added